+++ lib/request_handler/query_parser.rb
@@ -1,0 +1,23 @@
+# frozen_string_literal: true
+
+require 'request_handler/schema_parser'
+require 'request_handler/error'
+module RequestHandler
+  class QueryParser < SchemaParser
+    RESERVED_KEYS = %w(fields filter include page sort)
+
+    def initialize(params:, schema:, schema_options: {})
+      super(schema: schema, schema_options: schema_options)
+      @query = params.clone
+      RESERVED_KEYS.each { |key| query.delete(key) }
+    end
+
+    def run
+      validate_schema(query)
+    end
+
+    private
+
+    attr_reader :query
+  end
+end